pyproject.toml
[tool.poetry]
name = "amicia"
version = "0.1.0"
description = "Amicia search"
authors = ["Eugen Ciur <eugen@papermerge.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
xapianpy = "^1.4.22.post2305071405"
typer = "^0.9.0"
pydantic = "^2.0.3"
pytest = "^7.4.0"

[tool.poetry.scripts]
ami-cli = 'amicia.main:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
